refactor(UserToken): move token issuer to helpers folder
- refactor user token generator to be an helper function
- change import locations on files required
- adjust token verification to check for expired token
- rebased off last two commits

[Finishes #161786647]
