+++ server/src/controllers/UserController.js
@@ -1,7 +1,7 @@
 import bcrypt from 'bcrypt';
 import dotenv from 'dotenv';
 import jwt from 'jsonwebtoken';
-import userToken from '../middlewares/token';
+import userToken from '../helpers/token';
 import Models from '../db/models';
 import template from '../helpers/sendMail/templates';
 import sendMail from '../helpers/sendMail/sendMail';
